backend/requirements.txt
# Swasthya Saathi Backend

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
motor>=3.3.0
pymongo>=4.6.0
pytesseract>=0.3.10
Pillow>=10.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
python-dotenv>=1.0.0

